Refactor: extract ensure_async_token_provider into pyrit.auth.azure_auth
Move _ensure_async_token_provider from openai_target.py and
azure_content_filter_scorer.py into a shared public helper in
pyrit/auth/azure_auth.py. Uses the more robust implementation
(inspect.iscoroutinefunction + isawaitable handling) and logs at
DEBUG level.
