+++ pyrit/auth/__init__.py
@@ -10,6 +10,7 @@ from pyrit.auth.azure_auth import (
     AsyncTokenProviderCredential,
     AzureAuth,
     TokenProviderCredential,
+    ensure_async_token_provider,
     get_azure_async_token_provider,
     get_azure_openai_auth,
     get_azure_token_provider,
@@ -27,6 +28,7 @@ __all__ = [
     "CopilotAuthenticator",
     "ManualCopilotAuthenticator",
     "TokenProviderCredential",
+    "ensure_async_token_provider",
     "get_azure_token_provider",
     "get_azure_async_token_provider",
     "get_default_azure_scope",
