+++ pyrit/auth/azure_auth.py
@@ -121,6 +121,49 @@ class AsyncTokenProviderCredential:
         await self.close()
 
 
+def ensure_async_token_provider(
+    api_key: str | Callable[[], str | Awaitable[str]] | None,
+) -> str | Callable[[], Awaitable[str]] | None:
+    """
+    Ensure the api_key is either a string or an async callable.
+
+    If a synchronous callable token provider is provided, it's automatically wrapped
+    in an async function to make it compatible with async Azure SDK clients.
+
+    Args:
+        api_key: Either a string API key or a callable that returns a token (sync or async).
+
+    Returns:
+        Either a string API key or an async callable that returns a token.
+    """
+    if api_key is None or isinstance(api_key, str) or not callable(api_key):
+        return api_key
+
+    # Check if the callable is already async
+    if inspect.iscoroutinefunction(api_key):
+        return api_key
+
+    # Wrap synchronous token provider in async function
+    logger.debug(
+        "Detected synchronous token provider."
+        " Automatically wrapping in async function for compatibility with async client."
+    )
+
+    async def async_token_provider() -> str:
+        """
+        Async wrapper for synchronous token provider.
+
+        Returns:
+            str: The token string from the synchronous provider.
+        """
+        result = api_key()
+        if inspect.isawaitable(result):
+            return await result
+        return result
+
+    return async_token_provider
+
+
 class AzureAuth(Authenticator):
     """
     Azure CLI Authentication.
